create a common Json instance
