--- bridge/src/main/java/com/bitwarden/bridge/util/InstantSerializer.kt
@@ -1,22 +1,0 @@
-package com.bitwarden.bridge.util
-
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-import java.time.Instant
-
-/**
- * A simple serializer for serializing [Instant].
- */
-internal object InstantSerializer : KSerializer<Instant> {
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("java.time.Instant", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: Instant) =
-        encoder.encodeString(value.toString())
-
-    override fun deserialize(decoder: Decoder): Instant = Instant.parse(decoder.decodeString())
-}
